fix: feedback from code review
